HighMemDxe: rework as a FdtBusDxe client
Fixes to FdtBusDxe/DtIo:
- Fix bug in parsing property where prop len and cell count was confused
- Parse out device_type as DtIo.DeviceType.

Testing of HighMemDxe with RV: HighMemDxe is actually useless
to current RiscVVirt because PrePi code adds all ranges. To test,
do this:

--- a/OvmfPkg/RiscVVirt/Sec/Memory.c
+++ b/OvmfPkg/RiscVVirt/Sec/Memory.c
@@ -228,7 +228,7 @@ MemoryPeimInitialization (
             MmodeResvSize
             );
         } else {
-          AddMemoryBaseSizeHob (CurBase, CurSize);
+          //AddMemoryBaseSizeHob (CurBase, CurSize);
         }

You will now see these messages if you boot with 4gib RAM:
ProcessMemoryRange: Add System RAM @ 0x100000000 - 0x1BFFFFFFF
ProcessMemoryRange: nothing to do for 0x80000000-0xBFFFFFFF

Another outstanding problem is that driver model devices are
only connected when the handle is published or when the devices
are needed. HighMemDxe loads after FdtBusDxe, so nothing happens
until the ConnectAll done by BDS when no bootable boot options exist.
This isn't a problem for VirtioFdtBusDxe because all of those
devices really are optional for boot, but HighMemDxe ranges aren't
optional - they always need to exist.

One way is to force Bds to look up and connect the path in
PlatformBootManagerLib. This is not great because you need to know
the device path and there could be multiple ranges depending how much
RAM you boot with. Or add an API to quickly iterate over handles that
match a certain device type or compatible string - FdtClientDxe-style.
Another way is to treat some nodes specially
in FdtBusDxe and hook gEfiEndOfDxeEventGroupGuid to force their
connection. Another way is to make HighMemDxe do something to kick-off
the binding protocol check once installed.

Signed-off-by: Andrei Warkentin <andrei.warkentin@intel.com>