FdtBusDxe: translate addresses when looking/parsing up reg range.
- Today only identity translations are supported.
- EFI_DT_REG embedds DtIo pointer for device managing the bus address.
- An EFI_DT_REG BusDtIo == NULL means addresses are CPU real.
- Enforce fully translated (CPU real) address to be valid.

The intention is that driver binding will be able to set some
accessors used by PollReg/ReadReg/WriteReg/CopyReg. This will allow,
for example, a child device to perform register I/O without having
to know if the EFI_DT_REG encodes a CPU or a bus-specific address
only valid in the context of the parent device driver.

Signed-off-by: Andrei Warkentin <andrei.warkentin@intel.com>