Fix builds using SseUefiPkg.dsc due to missing deps.
AArch64 build still broken due to missing __ashlti3 dependency
in ArmPkg/Library/CompilerIntrinsicsLib. This is due to the
way we're defining EFI_DT_BUS_ADDRESS/EFI_DT_SIZE. This is
of course fixable in a number of ways (implementing __ashlti3,
narrowing down as a config option, to reworking EFI_DT_REG manipulation entirely,
which would be kind of nasty).

Signed-off-by: Andrei Warkentin <andrei.warkentin@intel.com>