Introduce DtIo.GetStringIndex
This is useful to look up other properties indexed by name, e.g.:
foo = < value1 value2 value3 >
foo-names = "index1", "index2", "index3"

DtIo->GetStringIndex (DtIo, "foo-names", "index3", &Index) will
set Index to 2, which you can then feed into another GetXXX or
ParseProp.

It's true that the "reg-names" and similar properties aren't
in the DT spec, but the philosophy behind the FdtBusDxe
interface is to minimize boilerplate code in drivers, and
"reg-names" is a common pattern adopted everywhere (since it
is adopted in Linux and U-Boot).

This is portion of a reworked https://github.com/intel-sandbox/SseUefiPkg/pull/60

Signed-off-by: Evan Chai <evan.chai@intel.com>
Signed-off-by: Andrei Warkentin <andrei.warkentin@intel.com>