FdtBusDxe: improve handling of test devices and add sample device.
It used to be that every test device (DT_DEVICE_TEST) was
a unit test. DT_DEVICE_TEST was used both to identify which
Devicetree the controller belonged to, and to signal
that FdtBusDxe should bind to the controller and run tests...

To play around with the sample driver code, add a sample device
to the test Devicetree. But it's obviously not a unit test device.
So now DT_DEVICE_TEST is split into DT_DEVICE_TEST_UNIT and DT_DEVICE_TEST.
DT_DEVICE_TEST_UNIT is set by the presence of uefi,unit-test-device
and inherited by children (so all unit tests are now moved
under a unit-test-devices node). DT_DEVICE_TEST_RAN is renamed
to DT_DEVICE_TEST_UNIT_RAN.

Signed-off-by: Andrei Warkentin <andrei.warkentin@intel.com>