PciHostBridgeFdtDxe: improve IO space handling in GCD
This is different from how the MdeModulePkg PciHostBridgeDxe does it
in code, but not different from how it does it in practice, seeing
that PciHostBridgeDxe is never truly aware of the IO->MMIO
translation offset (which is reported OOB via PcdPciIoTranslation).

The IO space is entirely fake and there's very little
reason to use the translated MMIO (host) addresses for
GCD manipulation, at the very least because frequently
the IO space is sized at 16 or 32 bits in the CPU HOB, but
also because the CpuIo2 protocol expects the addresses
to effectively be in this [0..0xffff(ffff)] range, so
keeping them in the GCD that way is at least consistent.

HostBridge AllocateResource thus *also* is only
passed device addresses, not translated addreses, for PCI I/O.

For the sake of argument, if FdtBusPkg is ever used with x86,
then you're looking at PCI I/O to CPU I/O translation with
an offset of 0, so using the PCI I/O address for manipulating
the I/O GCD is perfectly fine.

Signed-off-by: Andrei Warkentin <andrei.warkentin@intel.com>