FdtBusPkg: introduce DtInterrupt protocol
This can be implemented by a DT interrupt controller driver.

FbpInterruptUtilsLib can be used by a driver that needs
interrupt support (e.g. a DT timer driver). Functionality
is not part of FdtBusDxe because in practice it will only be
used by one driver - the timer driver.

FbpInterruptGet also supports an interrupt nexus.

Signed-off-by: Andrei Warkentin <andrei.warkentin@intel.com>