PciHostBridgeFdtDxe: support fully synthetic PCIe MMIO/IO apertures in the no-reconfiguration case.
In some scenarios, actual access to a PCIe device could be indirect,
i.e. not a 1:1 mapping between a BAR address and a CPU address.

UEFI expects the apertures to be visible in CPU space, which I suppose
isn't really a problem for well-written driver code (that uses
the PCI IO protocol for all I/O). However, PciHostBridgeFdtDxe
and other similar code relies on GCD functions to allocate/validate
ranges (see AllocateResource in HostBridge.c), so supporting
full configuration in this indirect aperture case would involve
an internal allocator (as the "indirect" part implies the PCI MMIO/IO
addresses are not part of the GCD!).

The good news is that, so long as we're keeping existing configuration
(fdtbuspkg,pci-keep-config), we can limp along. This is useful
in the case of UEFI userspace emulation (exposing devices via VFIO)
or other non-standard approaches.

The new ResTracked field PCI_RES_NODE tracks whether the resource
is tracked in the GCD.

Signed-off-by: Andrei Warkentin <andrei.warkentin@intel.com>